Finished the Recipe Pages and Reduced Clutter
To Reduce clutte on the index.html pages I got rid
of all of the images of the dishes along with there
titles.

For each of the recipes I go all of there ingreditents
as well as a brief description of the recipes and what
they are like. I also added the images to the page and a
source part so that people could go to where the recipe
is from to read the full directions if they wanted to.