Messed with Odin recipe page a bit.
I got the recipe to align in the center, and changed in
the background color. I have a really cool idea for how
to do a cool designed recipe page in the future, but I need to
learn more css so I can get the layout done. I have decided to wait
to look into the box method until after I get there in Odin.